vmwatcher: paid users only get 2x what free users get, move plan  related things to plan.go
